pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-crm-reinforcement"
version = "0.2.0"
description = "DataScan CRM Reinforcement Learning Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
    "requests",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
# This 'find' method is what the error message is asking for
packages = {find = {where = ["."], include = ["server*"]}}